Move AddressBooks & Contacts from MailData to MainViewModel
