+++ app/src/main/java/com/infomaniak/mail/ui/main/MainActivity.kt
@@ -31,7 +31,6 @@ import androidx.navigation.NavDestination
 import androidx.navigation.fragment.NavHostFragment
 import com.infomaniak.lib.core.utils.LiveDataNetworkStatus
 import com.infomaniak.mail.R
-import com.infomaniak.mail.data.MailData
 import com.infomaniak.mail.databinding.ActivityMainBinding
 import com.infomaniak.mail.ui.main.menu.MenuDrawerFragment
 import com.infomaniak.mail.utils.UiUtils
@@ -79,7 +78,7 @@ class MainActivity : ThemedActivity() {
         setupNavController()
         setupMenuDrawerCallbacks()
 
-        MailData.loadAddressBooksAndContacts()
+        mainViewModel.loadAddressBooksAndContacts()
     }
 
     private fun listenToNetworkStatus() {
